Move arithmetic ops to module.
